fix(clubs): apply code review comments and move enum to interface
