+++ packages/interface/src/api/clubs/endpoints/apiClb001.ts
@@ -1,6 +1,8 @@
 import { HttpStatusCode } from "axios";
 import { z } from "zod";
 
+import { clubTypeEnum } from "@sparcs-clubs/interface/common/enum/clubs";
+
 /**
  * @version v0.1
  * @description 전체 동아리 목록을 가져옵니다
@@ -25,7 +27,7 @@ const responseBodyMap = {
           .object({
             id: z.number().int().positive(),
             name: z.string().max(20),
-            type: z.string().max(10), // 동아리 유형(정동아리 | 가동아리 | 상임동아리)
+            type: clubTypeEnum, // 동아리 유형(정동아리 | 가동아리 | 상임동아리)
             characteristic: z.string().max(50), // 동아리 소개
             representative: z.string().max(20), // 동아리 대표
             advisor: z.string().max(20).nullable(), // 동아리 지도교수
