+++ packages/web/src/features/clubs/services/getClubsList.ts
@@ -7,7 +7,7 @@ import {
   UnexpectedAPIResponseError,
 } from "@sparcs-clubs/web/lib/axios";
 import apiClb001 from "@sparcs-clubs/interface/api/clubs/endpoints/apiClb001";
-import mockupClubList from "@sparcs-clubs/web/features/clubs/services/mockupClubList";
+import mockupData from "@sparcs-clubs/web/features/clubs/services/mockupClubData";
 
 // ISuccessResponseType.clubs는 clubs.types.ts의 ClubInfo와 대응됩니다.
 // ClubType의 타입이 불일치하기 때문에, 사용 이전에 ClubType Enum에 매칭하는 과정이 필요합니다.
@@ -31,8 +31,7 @@ export const getClubsList = () =>
 
 defineAxiosMock(mock => {
   mock.onGet(apiClb001.url()).reply(() => {
-    const dummy: z.infer<(typeof apiClb001.responseBodyMap)[200]> =
-      mockupClubList;
+    const dummy: z.infer<(typeof apiClb001.responseBodyMap)[200]> = mockupData;
     return [200, dummy];
   });
 });
